requirements.txt
bcrypt
psycopg2
pycryptodome
flask
flask-cors
flask-jwt-extended
python-dotenv